+++ src/app/componentes/hard-soft-skills/hard-soft-skills.component.ts
@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Frases } from 'src/app/model/frases';
 import { Hardsskills } from 'src/app/model/hardsskills';
 import { NuevoUsuario } from 'src/app/model/nuevo-usuario';
 import { AuthService } from 'src/app/service/auth.service';
+import { EnvioUsuarioIdService } from 'src/app/service/envio-usuario-id.service';
 import { SFrasesService } from 'src/app/service/s-frases.service';
 import { SHardSSkillsService } from 'src/app/service/s-hard-sskills.service';
 import { TokenService } from 'src/app/service/token.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-hard-soft-skills',
@@ -28,9 +31,11 @@ export class HardSoftSkillsComponent implements OnInit {
 
   constructor(
     private sFrases: SFrasesService,
+    private router: Router,
     private sHardSSkillsService: SHardSSkillsService,
     private tokenService: TokenService,
-    private authService: AuthService
+    private authService: AuthService,
+    public envioUsuarioIdService: EnvioUsuarioIdService,
   ) {
   }
 
@@ -39,6 +44,12 @@ export class HardSoftSkillsComponent implements OnInit {
   ngOnInit(): void {
     this.token();
     this.cargarUsuarioId();
+    this.envioUsuarioIdService.cargadorUsuarioId.subscribe(
+      data=>{
+        this.cargarHardsSkills(data.data);
+        this.cargarFraseUsuarioId(data.data);
+      }
+    )
   }
 
   token() {
@@ -54,6 +65,7 @@ export class HardSoftSkillsComponent implements OnInit {
         this.hardsskills = JSON.parse(JSON.stringify(data));
       })
   }
+  
   getUsuarioId() {
     if (this.usuarioId == undefined) {
       this.usuarioId = 1
@@ -76,7 +88,7 @@ export class HardSoftSkillsComponent implements OnInit {
             this.frase = x.frases;
             this.autor = x.autor;
             this.idFrases = x.id;
-            console.log("frases: "+x.frases+" autor: "+x.autor+ " id: "+ x.id+"  seccionID: "+ x.seccionId+"  usuarioId: "+x.usuarioId)
+            // console.log("frases: "+x.frases+" autor: "+x.autor+ " id: "+ x.id+"  seccionID: "+ x.seccionId+"  usuarioId: "+x.usuarioId)
           }
         })
       })
@@ -100,17 +112,44 @@ export class HardSoftSkillsComponent implements OnInit {
       })
   }
 
-  delete(id?: number) {
+delete(nombres: string,id?: number) {
     if (id != undefined) {
-      this.sHardSSkillsService.delete(id).subscribe(
-        {
-          next: () => { this.cargarHardsSkills(this.usuarioId) },
-          error: () => { alert("No se pudo eliminar") },
-          complete: () => { console.info('complete') }
+      Swal.fire({
+        title: 'Quieres borrar '+ nombres + '?',
+        text: "Esto será irreversible! ",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, borrar '+nombres + '!'
+      })
+        .then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Borrado!',
+              nombres+ ' ha sido eliminado.',
+              'success'
+            )
+            this.sHardSSkillsService.delete(id).subscribe(
+              {
+                next: () => {
+                  { this.cargarHardsSkills(this.usuarioId) }
+                  this.router.navigate([''])
+                },
+                error: (e) => {
+                  Swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'No se pudo eliminar!',
+                    footer: 'error: '+e
+                  })
+                },
+                complete: () => { console.log('complete') }
+              })
+          }
         })
     }
   }
-
 }
